decimal float

+++ incomecalculation/models.py
@@ -28,7 +28,7 @@ class IncomeSettings(TimeStampModel):
 
 class ManagerialIncome(TimeStampModel):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    index = models.DecimalField(decimal_places=0, max_digits=20)
+    index = models.FloatField()
     total_score = models.DecimalField(decimal_places=0, max_digits=20)
     history_income = models.DecimalField(decimal_places=2, max_digits=20)
     position_income = models.DecimalField(decimal_places=2, max_digits=20)
@@ -47,8 +47,8 @@ class ManagerialIncome(TimeStampModel):
 
 class ManagerialFixIncome(TimeStampModel):
     user = models.ForeignKey(User, on_delete=models.CASCADE)
-    history_income_percentage = models.DecimalField(decimal_places=2, max_digits=20)
-    position_income_percentage = models.DecimalField(decimal_places=2, max_digits=20)
+    history_income_percentage = models.FloatField()
+    position_income_percentage = models.FloatField()
 
     class Meta:
         db_table = 'managerial_fix_incomes'
